Create RandDieGen
Made a couple of changes to your code, should work properly now but test it and tweak it to your needs brother, also make this private so your professor doesn't find it and gives you an F for cheating or getting outside help. 

Just a pointer in case you have to write code again, which you will need if you work on electrical systems or anything that needs to be controlled by a computer, make sure to document and make any comments to why you changed certain snippets of code, that way it will be easier to keep maintence up on your code and you 'll already know where you are and how to fix your code. Also you're pretty much writing an algorithm when you document your thought process. its just good practice.